requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0 
djangorestframework-gis==1.0
psycopg==3.1.8
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0    
tzdata==2023.3